refactor: use deserialiser from log::Level, and mention example TOML
use the serde functionality of the log crate. this will make the error
message print the accepted strings if an incorrect string is provided.

    unknown variant `asd`, expected one of `ERROR`, `WARN`, `INFO`, `DEBUG`, `TRACE`

however, if you use a wrong type (like bool), it will just say
"wanted string or table".

to try and direct to docs, we link to the example toml in github.
alternatively, this could link to https://lychee.cli.rs/guides/config/
instead?

the table format from log::Level is a bit mysterious and could be
a source for confusion, but idk how to prevent that - can you somehow
assert that what is being deserialiesd is a string?

related to https://www.github.com/lycheeverse/lychee/issues/1903