refactor: check glob validity when parsing input source arguments (#1869)
* refactor: check glob validity when parsing input source arguments

now, when providing an input with glob characters but an invalid glob
pattern, this will cause a "cannot parse inputs" error rather than
erroring out later during link checking. this makes use of the
existing ErrorKind::InvalidGlobPattern error case.

for example, a glob with unbalanced brackets is invalid and now looks
like this:
```console
$ cargo run -- 'noasdjfi/['
Error: Cannot parse inputs from arguments

Caused by:
    0: UNIX glob pattern is invalid
    1: Pattern syntax error near position 9: invalid range pattern
```

this PR was prompted by findings in
https://github.com/lycheeverse/lychee/pull/1864#issuecomment-3385187203

* add one test case

* review: change FsGlob field to be a glob::Pattern

unfortunately, the `glob_with` function (which walks the dirs and
performs the globbing) does _not_ support taking a Pattern as input, so
we have to convert it back to a string at that point. this is a bit
weird, but i think it's still a bit nicer.

* refactor: move InputSource parsing into InputSource::new

this parsing logic is also changed to use early return so there is less
nesting

* fix lint

Co-authored-by: Thomas Zahner <thomas.zahner@protonmail.ch>

* use unwrap to ensure both Ok, and remove test helper function

---------

Co-authored-by: Thomas Zahner <thomas.zahner@protonmail.ch>