refactor: use InputResolver to implement Input::get_sources  (#1880)
previously, there was a lot of duplicated code between
InputResolver::resolve_input and Input::get_sources, where
resolve_input is used for main link checking and get_sources is used for
dumping inputs.

after this PR, Input::get_sources is implemented in terms of
InputResolver::resolve. however, in doing so, we have to loosen the
lifetime constraints on the `&Input` argument of resolve, so there's a
lot of changes in resolve_input to effectively _transpose_ the logic.
that is, previously it matched on the input source within a single
try_stream!, and now it matches on the input source first then chooses
between multiple try_streams. however, because try_stream constructs a
new anonymous type every use, we then need to change the type to pin box
dyn send.

the transposing of resolve_input does open the possibility of returning a `Result<Box<dyn Stream ...>>`
to represent input source failures, as desired in
https://www.github.com/lycheeverse/lychee/pull/1864#issuecomment-3389278177