Fix parsing larger HTML blocks in MDX files
Previously, each HTML line inside a Markdown file would be parsed
independently. This worked reasonably well, but in the case where
individual lines are not valid HTML, we would skip link checking.

The fix is to accumulate an HTML block until it gets closed and
only _then_ parsing it.

Fixes #1923