feat: support relative --root-dir (#1912)
* feat: support relative --root-dir

done by canonicalizing the root-dir at the point where
the is_absolute check was previously done. this is the
quick-n-fast way of making this change, as it maintains
all following program invariants - the root-dir is always
absolute past that point.

the more sophisticated approach would allow relative
root-dirs to pass unimpeded and it should all "just work"
down the line. but i didnt want to take the risk

a side-effect of this change is that a non-existing root-dir
is now a CLI error. previously, this would cause link checking
errors when it came time to check relative links which used
the non-existing root-dir.

* add test

* fix test oops

* do not canonicalize if base-url is given. however,

this creates a hole where a non-absolute root-dir may slip through if a
base-url is specified. this is PROBABLY fine, since it gets joined onto
base url anyway

* cargo_bin_cmd