Fully hide progress bar if any input is `stdin`
In https://github.com/lycheeverse/lychee/pull/1914, work was done to hide the
progress bar on `stdin`. The simple solution of removing enable_steady_tick for
the progress bar. This worked well for avoiding cluttering the output while the
user typed their input.

However, the initial progress bar still gets shown, which is not ideal. This
commit slightly imporves the behavior by waiting for all inputs before showing
the progress bar. We do this by iterating over the inputs and checking if any
of them is `stdin`.

This way, the progress bar does no longer get shown. To start the processing,
the user has to press Ctrl^D, which is customs for other Unix tools as well.
Other behavior remains unchanged.