Extract URLs from CSS
In #1485 it was requested to extract URLs from CSS files and inline
styles.

While lychee proper is generally becoming more conservative about the
file types we support, the places where links can occur in CSS are
pleasantly limited to `url()` directives (with the exception of inline
URLs in `font-face` and `@import` rules, which we conveniently ignore
for now).

As such, integrating CSS support into lychee can be done by means of a
simple extractor utilizing a bunch of regular expressions. It's not
quite as exhaustive as a proper CSS parser (such as Servo's
rust-cssparser crate), but on the plus side, we don't need to add
additional dependencies.

Fixes #1485