Spend way too much time on validation
Okay, this is slightly annoying. There are a lot of edge-cases with
input validation (who would have guessed?). I hope to have reached an
agreeable middle-ground by dividing validation into an early-stage if
the inputs are clearly file-paths and a deferred late-stage where they
are not (but might be). Not sure if we can do any better at this point.

This also means that one test for an non-existent file falls into the
early-validation category now and returns a better (I think) error
message. That's lovely.